Removed use of participant_events view for max sequential id lookup
