+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/h2/H2StorageBackend.scala
@@ -6,7 +6,6 @@ package com.daml.platform.store.backend.h2
 import java.sql.Connection
 import java.time.Instant
 import anorm.{Row, SQL, SimpleSql}
-import anorm.SqlParser.get
 import com.daml.ledger.offset.Offset
 import com.daml.lf.data.Ref
 import com.daml.logging.LoggingContext
@@ -115,23 +114,6 @@ private[backend] object H2StorageBackend
 
   override def insertBatch(connection: Connection, batch: AppendOnlySchema.Batch): Unit =
     H2Schema.schema.executeUpdate(batch, connection)
-
-  def maxEventSequentialIdOfAnObservableEvent(
-      offset: Offset
-  )(connection: Connection): Option[Long] = {
-    import com.daml.platform.store.Conversions.OffsetToStatement
-    SQL"""
-SELECT max_esi FROM (
-  (SELECT max(event_sequential_id) AS max_esi FROM participant_events_consuming_exercise WHERE event_offset <= $offset GROUP BY event_offset ORDER BY event_offset DESC FETCH NEXT 1 ROW ONLY)
-  UNION ALL
-  (SELECT max(event_sequential_id) AS max_esi FROM participant_events_non_consuming_exercise WHERE event_offset <= $offset GROUP BY event_offset ORDER BY event_offset DESC FETCH NEXT 1 ROW ONLY)
-  UNION ALL
-  (SELECT max(event_sequential_id) AS max_esi FROM participant_events_create WHERE event_offset <= $offset GROUP BY event_offset ORDER BY event_offset DESC FETCH NEXT 1 ROW ONLY)
-) AS t
-ORDER BY max_esi DESC
-FETCH NEXT 1 ROW ONLY;
-       """.as(get[Long](1).singleOpt)(connection)
-  }
 
   object H2QueryStrategy extends QueryStrategy {
